Added a new script to compare variables between two graphwise-styleguide versions
## What
Added a new script that compares the CSS variables of two different graphwise-styleguide versions.

## Why
To simplify the process of finding and fixing removed, renamed, or value-changed variables. This helps detect unwanted changes early and makes it easier to fix them in time.

## How
The new script downloads a specific version of graphwise-styleguide from npm and compares its variables with those from the version currently used in the project. It then generates an HTML report containing a table of all variables that were removed or whose values have changed. The report includes filters to reduce displayed changes.